Module2/api/requirements.txt
fastapi
uvicorn
mysql-connector-python
pika
psycopg2-binary